Added WAV streaming output for apu to debug it
- Removed the mixer logic and simplified to positive only output
- Removed old test roms and added back the ones that helped
- Started to write an NSF loader since those don't use PPU

APU starts to output data on the same frame that nestest crashes.
Only audio code execution is visible in the debugger.
Will switch to the old blargg tests that don't fail the same way

+++ source/nes/apu/apu.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "../common.h"
+#include "../../common.h"
 #include "envelope.h"
 #include "pulseChannel.h"
 #include "triangleChannel.h"
@@ -49,7 +49,7 @@ public:
 
     // Does the mixdown of all the channels at the current moment in time
     // According to https://www.nesdev.org/wiki/APU_Mixer has a range of 0.0 - 1.0
-    float getOutput();
+    uint32 getOutput();
     
     PulseChannel pulse1;
     PulseChannel pulse2;
